+++ crates/fuel-core/src/service/genesis.rs
@@ -10,6 +10,7 @@ use crate::{
     service::config::Config,
 };
 use fuel_core_chain_config::GenesisCommitment;
+use fuel_core_services::StateWatcher;
 use fuel_core_storage::{
     iter::IteratorOverTable,
     tables::{
@@ -54,19 +55,16 @@ mod importer;
 mod task_manager;
 
 pub use exporter::Exporter;
-use tokio_util::sync::CancellationToken;
 
 use self::importer::SnapshotImporter;
 
 /// Performs the importing of the genesis block from the snapshot.
 pub async fn execute_genesis_block(
+    watcher: StateWatcher,
     config: &Config,
     db: &CombinedDatabase,
 ) -> anyhow::Result<UncommittedImportResult<Changes>> {
-    // TODO: tie this with a SIGNAL for resumability
-    let cancel = CancellationToken::new();
-    SnapshotImporter::import(db.clone(), config.snapshot_reader.clone(), cancel.clone())
-        .await?;
+    SnapshotImporter::import(db.clone(), config.snapshot_reader.clone(), watcher).await?;
 
     let genesis_progress_on_chain: Vec<String> = db
         .on_chain()
@@ -139,7 +137,7 @@ pub async fn execute_and_commit_genesis_block(
     config: &Config,
     db: &CombinedDatabase,
 ) -> anyhow::Result<()> {
-    let result = execute_genesis_block(config, db).await?;
+    let result = execute_genesis_block(StateWatcher::default(), config, db).await?;
     let importer = fuel_core_importer::Importer::new(
         config
             .snapshot_reader