+++ crates/fuel-core/src/service/genesis/exporter.rs
@@ -18,6 +18,7 @@ use fuel_core_chain_config::{
     StateConfigBuilder,
     TableEntry,
 };
+use fuel_core_services::State;
 use fuel_core_storage::{
     blueprint::BlueprintInspect,
     iter::IterDirection,
@@ -34,8 +35,7 @@ use fuel_core_storage::{
 };
 use fuel_core_types::fuel_types::ContractId;
 use itertools::Itertools;
-
-use tokio_util::sync::CancellationToken;
+use tokio::sync::watch;
 
 use super::task_manager::TaskManager;
 
@@ -57,12 +57,15 @@ where
         writer: Fun,
         group_size: usize,
     ) -> Self {
+        // TODO: Support graceful shutdown during the exporting of the snapshot.
+        //  https://github.com/FuelLabs/fuel-core/issues/1828
+        let (_, receiver) = watch::channel(State::Started);
         Self {
             db,
             prev_chain_config,
             writer,
             group_size,
-            task_manager: TaskManager::new(CancellationToken::new()),
+            task_manager: TaskManager::new(receiver.into()),
         }
     }